pyproject.toml
[tool.poetry]
name = "mycharges"
version = "0.1.0"
description = "Telegram bot application to report on Investec account and card charges"
authors = ["Tai Lucas <tglucas@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
app = "app.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
awscli = "^1.27.69"
sentry-sdk = "^1.15.0"
simplejson = "^3.18.3"
pylib = {path = "pylib"}
python-telegram-bot = "^20.2"
emoji = "^2.2.0"
boto3 = "^1.28.19"
pymongo = {extras = ["srv"], version = "^4.4.1"}
investec-api-python = {path = "investec-api-python"}
lets-plot = "^3.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
aiosqlite = "^0.19.0"
pycryptodome = "^3.18.0"
influxdb-client = "^1.37.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
